classec/requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.9
PyMySQL==0.7.9
SQLAlchemy==1.1.4
Werkzeug==0.11.11
wheel==0.24.0

#Extra
bs4==0.0.1
Flask-SQLALchemy==2.1
boto3==1.4.1
requests==2.12.3
http://cdn.mysql.com//Downloads/Connector-Python/mysql-connector-python-2.0.5.tar.gz
